Unclutered Statsd TCP and UDP classes
Pulled out shared interface into separate file.
Pulled out address resolution.
