+++ src/StatsdClient/StatsdUDP.cs
@@ -6,19 +6,12 @@ using System.Text;
 
 namespace StatsdClient
 {
-    public interface IStatsdClient : IDisposable
-    {
-        void Send(string command);
-    }
-
-    public class StatsdUDPClient : IStatsdClient
+    public class StatsdUDPClient : Address, IStatsdClient
     {
         public IPEndPoint IPEndpoint { get; private set; }
 
         private readonly int _maxUdpPacketSizeBytes;
         private readonly Socket _clientSocket;
-        private readonly string _name;
-        private readonly int _port;
 
         /// <summary>
         /// Creates a new StatsdUDP class for lower level access to statsd.
@@ -26,38 +19,13 @@ namespace StatsdClient
         /// <param name="name">Hostname or IP (v4) address of the statsd server.</param>
         /// <param name="port">Port of the statsd server. Default is 8125.</param>
         /// <param name="maxUdpPacketSizeBytes">Max packet size, in bytes. This is useful to tweak if your MTU size is different than normal. Set to 0 for no limit. Default is MetricsConfig.DefaultStatsdMaxUDPPacketSize.</param>
-        /// <param name="useTcpProtocol">Sets the protocol type on Socket.</param>
         public StatsdUDPClient(string name, int port = 8125, int maxUdpPacketSizeBytes = MetricsConfig.DefaultStatsdMaxUDPPacketSize)
         {
-            _name = name;
-            _port = port;
             _maxUdpPacketSizeBytes = maxUdpPacketSizeBytes;
 
             _clientSocket = new Socket(AddressFamily.InterNetwork, SocketType.Dgram, ProtocolType.Udp);
 
-            var ipAddress = GetIpv4Address(name);
-            IPEndpoint = new IPEndPoint(ipAddress, _port);
-        }
-
-        private IPAddress GetIpv4Address(string name)
-        {
-            IPAddress ipAddress;
-            var isValidIpAddress = IPAddress.TryParse(name, out ipAddress);
-
-            if (!isValidIpAddress)
-            {
-                ipAddress = GetIpFromHostname();
-            }
-
-            return ipAddress;
-        }
-
-        private IPAddress GetIpFromHostname()
-        {
-            var addressList = Dns.GetHostEntry(_name).AddressList;
-            var ipv4Addresses = addressList.Where(x => x.AddressFamily != AddressFamily.InterNetworkV6);
-
-            return ipv4Addresses.First();
+            IPEndpoint = new IPEndPoint(GetIpv4Address(name), port);
         }
 
         public void Send(string command)
