+++ src/Tests/App.config
@@ -3,7 +3,7 @@
 <configuration>
   <appSettings>
     <!--<add key="StatsdServerName" value="localhost"/>-->
-    <add key="StatsdServerName" value="192.168.1.13"/>
+    <add key="StatsdServerName" value="192.168.1.14"/>
     <!--<add key="StatsdServerPort" value="8125"/>-->
     <add key="StatsdServerPort" value="8094"/>
